Lihuen GPS module test script using NeoGPS lib.
The following script was implemented by C. Correa on 2023-03-03 to find a more robust and reliable way to gat a timestamp from the GPS module of Humedata. The previous version tried TinyGPS++ lib (CS, Feb 2023) but both HD Atlas and Lihuen failed to produce reliable timestamps for storage in the SD memories.