Battery level fixes
A number of changes to enhance and store battery level information. Includes new parameterization based on empirical data to convert from analog readings to Volts. Made new variables, and renamed others to make the code more intuitive and complete. Correct Volts are now stored in the SD cards of both Atlas and Lihuen.